Well...that was embarrassing...

+++ Makefile
@@ -3,10 +3,5 @@ all:
 	mv ./.build/debug/TikTakToe .
 
 clean:
-	if [ -f ./TikTakToe ]; then
-		rm ./TikTakToe
-	fi
-
-	if [ -d ./.build/ ]; then
-		rm -rf ./.build/
-	fi
+	if [ -f ./TikTakToe ]; then rm ./TikTakToe; fi
+	if [ -d ./.build/ ]; then rm -rf ./.build/; fi
